common/xmalloc_tlsf: change bitmaps to unsigned long instead of u32
Atomic operations are used on these bitmaps and atomics take a performance
hit on some architectures if they are not aligned to the architecture's
integer width (64-bit on x86_64, for example).  On other architectures,
misaligned atomics are simply not supported (RISC-V).

This commit changes these bitmaps to be "unsigned long" in order to
cause these fields to be automatically aligned to the integer width of
the architecture when compiled with strict alignment or on compiler
targets that default to strict alignment.

+++ xen/common/xmalloc_tlsf.c
@@ -91,10 +91,10 @@ struct bhdr {
 
 struct xmem_pool {
     /* First level bitmap (REAL_FLI bits) */
-    u32 fl_bitmap;
+    unsigned long fl_bitmap;
 
     /* Second level bitmap */
-    u32 sl_bitmap[REAL_FLI];
+    unsigned long sl_bitmap[REAL_FLI];
 
     /* Free lists */
     struct bhdr *matrix[REAL_FLI][MAX_SLI];
@@ -174,7 +174,7 @@ static inline void MAPPING_INSERT(unsigned long r, int *fl, int *sl)
 static inline struct bhdr *FIND_SUITABLE_BLOCK(struct xmem_pool *p, int *fl,
                                                int *sl)
 {
-    u32 tmp = p->sl_bitmap[*fl] & (~0u << *sl);
+    unsigned long tmp = p->sl_bitmap[*fl] & (~0u << *sl);
     struct bhdr *b = NULL;
 
     if ( tmp )
